Fixed a really pesky PDL data transfer bug.
It turned out that the internal function that determines whether
the piddle is a slice/mmap or not was always using a double for
its piddle, not the passed-in piddle's type. This led to datarefs
that were too long, and that led to all sorts of memory issues.
This set of corrections should fix that, though I'm not sure that
the test suite is fully up-to-date.