Removed DeviceReset from the module.
After getting a couple of reports of difficulties with DeviceReset
functionality, I decided that it is not actually a useful function
for a simple set of bindings. It should certainly appear in the full
driver API wrapper, but it serves only to complicate matters in the
context of these bindings.